requirements.txt
# Local RAG Requirements
# Based on popular approaches from HN discussion

# Embeddings
sentence-transformers>=2.2.0

# Vector Search
faiss-cpu>=1.7.4

# SQLite with FTS5 (built-in Python)
# No additional package needed

# Ollama client for local LLM
ollama>=0.3.0
openai>=1.30.0

# Document processing
pypdf>=4.0.0
python-docx>=1.1.0
markdown>=3.5.0

# Utilities
numpy>=1.24.0
tqdm>=4.66.0
rich>=13.7.0
psutil>=5.9.8
